Use decimal as we want to be able to track half syns as well

+++ Syns/Application.cs
@@ -22,7 +22,7 @@
             }
         }
 
-        public int TodaySyns()
+        public decimal TodaySyns()
         {
             return m_SynsStore.GetTodaySyns(m_LoggedUserName);
         }
